Upgrade CPANPLUS to 0.83_08
p4raw-id: //depot/perl@32218

+++ lib/CPANPLUS/Internals/Source.pm
@@ -820,7 +820,7 @@ sub _create_mod_tree {
                                             # 'foo-bar-baz-1.03.tar.gz'
                 description => $dslip_tree->{ $data[0] }->{'description'},
                 dslip       => $dslip,
-                _id         => $self->_id,  #id of this internals object
+                _id         => $self->_id,  # id of this internals object
         );
 
     } #for
@@ -1107,10 +1107,15 @@ sub _remove_custom_module_source {
     ### use uri => local, instead of the other way around
     my %files = reverse $self->__list_custom_module_sources;
     
-    my $file = $files{ $uri } or do {
-                    error(loc("No such custom source '%1'", $uri));
-                    return;
-                };
+    ### On VMS the case of key to %files can be either exact or lower case
+    ### XXX abstract this lookup out? --kane
+    my $file = $files{ $uri };
+    $file    = $files{ lc $uri } if !defined($file) && ON_VMS;
+
+    unless (defined $file) {
+        error(loc("No such custom source '%1'", $uri));
+        return;
+    };
                 
     1 while unlink $file;
  
@@ -1242,8 +1247,13 @@ sub __update_custom_module_source {
             return;                      
         };
 
+        ### On VMS the case of key to %files can be either exact or lower case
+        ### XXX abstract this lookup out? --kane
+        my $file = $files{ $remote };
+        $file    = $files{ lc $remote } if !defined ($file) && ON_VMS;
+
         ### return the local file we're supposed to use
-        $files{ $remote } or do {
+        $file or do {
             error(loc("Remote source '%1' unknown -- needs '%2' argument",
                       $remote, 'local'));
             return;
@@ -1275,11 +1285,17 @@ sub __update_custom_module_source {
         } else {
             msg(loc("No index file found at '%1', generating one",
                     $ff->uri), $verbose );
+            
+            ### ON VMS, if you are working with a UNIX file specification,
+            ### you need currently use the UNIX variants of the File::Spec.
+            my $ff_path = do {
+                my $file_class = 'File::Spec';
+                $file_class .= '::Unix' if ON_VMS;
+                $file_class->catdir( File::Spec::Unix->splitdir( $ff->path ) );
+            };      
 
             $self->__write_custom_module_index(
-                path    => File::Spec->catdir(
-                                File::Spec::Unix->splitdir( $ff->path )
-                            ),
+                path    => $ff_path,
                 to      => $local,
                 verbose => $verbose,
             ) or return;
@@ -1347,7 +1363,7 @@ sub __write_custom_module_index {
         ### make sure to remove the leading slash as well.
         my $copy = $File::Find::name;
         my $re   = quotemeta($path);        
-        $copy    =~ s|^$path[\\/]?||i;
+        $copy    =~ s|^$re[\\/]?||i;
         
         push @files, $copy;
         
@@ -1434,7 +1450,11 @@ Returns true on success, false on failure.
                 
                 ### and now add it to the modlue tree -- this MAY
                 ### override things of course
-                if( $self->module_tree( $mod->module ) ) {
+                if( my $old_mod = $self->module_tree( $mod->module ) ) {
+
+                    ### On VMS use the old module name to get the real case
+                    $mod->module( $old_mod->module ) if ON_VMS;
+
                     msg(loc("About to overwrite module tree entry for '%1' with '%2'",
                             $mod->module, $mod->package), $verbose);
                 }
